static library draft (it crashes in `NP_PingPeer`)
